Fix aircopy bypassing TX restrictions & refactor
And scrapes together a few more bytes.

+++ frequencies.c
@@ -223,7 +223,7 @@ int32_t TX_freq_check(const uint32_t Frequency)
 			break;
 
 		case F_LOCK_NONE:
-			for (uint8_t i = 0; i < ARRAY_SIZE(frequencyBandTable); i++)
+			for (uint32_t i = 0; i < ARRAY_SIZE(frequencyBandTable); i++)
 				if (Frequency >= frequencyBandTable[i].lower && Frequency < frequencyBandTable[i].upper)
 					return 0;
 			break;
